some corrections

+++ DummyProjectBAL/UserBAL.cs
@@ -61,16 +61,19 @@ namespace DummyProjectBAL
 
         #endregion
 
-
-
-        #region All
-
-        public Result GetUserDetailsByID(Int64 ID)
+        # region GetUserDetailsByID
+        public Result GetUserDetailsByID(Int32 ID)
         {
             UserDAL userDAL = new UserDAL();
             return userDAL.GetUserDetailsByID(ID);
         }
-   
+        #endregion
+
+
+        #region All
+
+
+
         public Result UserLookup(String TypeHeadKeyword)
         {
             UserDAL userDAL = new UserDAL();
